Move root comment collapsing to sliding comment component
