+++ src/features/comment/Comment.tsx
@@ -3,7 +3,7 @@ import { IonIcon, IonItem } from "@ionic/react";
 import { chevronDownOutline } from "ionicons/icons";
 import { CommentView } from "lemmy-js-client";
 import { css } from "@emotion/react";
-import React, { useContext, useEffect, useRef } from "react";
+import React, { useEffect, useRef } from "react";
 import Ago from "../labels/Ago";
 import { maxWidthCss } from "../shared/AppContent";
 import PersonLink from "../labels/links/PersonLink";
@@ -13,9 +13,6 @@ import AnimateHeight from "react-animate-height";
 import CommentContent from "./CommentContent";
 import useKeyPressed from "../../helpers/useKeyPressed";
 import SlidingNestedCommentVote from "../shared/sliding/SlidingNestedCommentVote";
-import { useDispatch } from "react-redux";
-import { toggleCommentCollapseState } from "./commentSlice";
-import { AppContext } from "../auth/AppContext";
 
 const rainbowColors = [
   "#FF0000", // Red
@@ -172,7 +169,7 @@ interface CommentProps {
 
   className?: string;
 
-  rootIndex: number;
+  rootIndex?: number;
 }
 
 export default function Comment({
@@ -189,24 +186,8 @@ export default function Comment({
   rootIndex,
 }: CommentProps) {
   const keyPressed = useKeyPressed();
-  const dispatch = useDispatch();
   // eslint-disable-next-line no-undef
   const commentRef = useRef<HTMLIonItemElement>(null);
-
-  const { activePage } = useContext(AppContext);
-
-  function collapseRootComment() {
-    const rootCommentId = +comment.comment.path.split(".")[1];
-
-    dispatch(toggleCommentCollapseState(rootCommentId));
-
-    if (!activePage || !("current" in activePage)) return;
-
-    activePage.current?.scrollToIndex({
-      index: rootIndex,
-      behavior: "smooth",
-    });
-  }
 
   useEffect(() => {
     if (highlightedCommentId !== comment.comment.id) return;
@@ -225,7 +206,7 @@ export default function Comment({
       <SlidingNestedCommentVote
         item={comment}
         className={className}
-        collapse={collapseRootComment}
+        rootIndex={rootIndex}
         collapsed={!!collapsed}
       >
         <CustomIonItem
